requirement.txt
pydantic
fastapi
bs4
uvicorn 
requests
numpy
pydantic-settings 
python-dotenv